move files up a folder
